requirements.txt
connexion[swagger-ui]==2.5.1
uWSGI==2.0.18
ruamel.yaml==0.16.5
GitPython==3.0.5